requirements.txt
alembic==1.17.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==3.2.2
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cmdstanpy==1.3.0
cohere==5.20.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.121.1
fastavro==1.12.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-PyMongo==3.0.1
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.1
fpdf==1.7.2
fsspec==2025.10.0
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.42.0
google-auth-httplib2==0.2.1
google-genai==1.46.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==24.0.0
h11==0.16.0
hf-xet==1.2.0
holidays==0.84
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.1.4
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Mako==1.3.10
Markdown==3.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.10.0
networkx==3.5
numpy==2.3.4
openai==2.7.2
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
plotly==6.3.1
prophet==1.2.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.15.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
stanio==0.5.1
starlette==0.49.3
streamlit==1.50.0
streamlit-camera-input-live==0.2.0
streamlit-option-menu==0.4.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.2
WTForms==3.2.1
